requirements.txt
# Core dependencies
torch>=2.0.0
transformers>=4.35.0
langchain>=0.2.0
langchain-community>=0.2.0
langchain-core>=0.2.0

# Vector Database
chromadb>=0.4.0
sentence-transformers>=2.2.0

# Multimodal models
Pillow>=10.0.0
opencv-python>=4.8.0

# Qwen model support
dashscope>=1.17.0

# Data processing
numpy>=1.24.0
pandas>=2.0.0
json5>=0.9.0

# Utilities
tqdm>=4.66.0
requests>=2.31.0
python-dotenv>=1.0.0
accelerate>=1.0.0

# Web Interface
streamlit>=1.28.0

